Removed MultipleEventsCutter/clickableSingle, original reason for having this was due to Mutex bug. Haven't encountered it yet after playing multiple games, only worry is that it was random when it was first encountered before release. Will keep close eye on this. Simplified AutoSizeText into one file. Moved theme directory.
