+++ app/src/androidTest/java/com/heyzeusv/solitaire/ui/TableauTest.kt
@@ -9,7 +9,7 @@ import androidx.compose.ui.unit.dp
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.heyzeusv.solitaire.R
 import com.heyzeusv.solitaire.ui.game.SolitaireTableau
-import com.heyzeusv.solitaire.ui.theme.SolitaireTheme
+import com.heyzeusv.solitaire.util.theme.SolitaireTheme
 import com.heyzeusv.solitaire.util.TestCards
 import com.heyzeusv.solitaire.util.onCard
 import com.heyzeusv.solitaire.util.onNodeWithConDescId
