--- app/src/main/java/com/heyzeusv/solitaire/ui/Modifier.kt
@@ -1,43 +1,0 @@
-package com.heyzeusv.solitaire.ui
-
-import androidx.compose.foundation.LocalIndication
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.interaction.MutableInteractionSource
-import androidx.compose.runtime.remember
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.composed
-import androidx.compose.ui.platform.debugInspectorInfo
-import androidx.compose.ui.semantics.Role
-
-/**
- *  Found here [https://al-e-shevelev.medium.com/how-to-prevent-multiple-clicks-in-android-jetpack-compose-8e62224c9c5e].
- *  Was having crashes that were very random and impossible to reproduce.
- *  kotlinx.coroutines.CompletionHandlerException: Exception in resume onCancellation handler for
- *  CancellableContinuation... Caused by: java.lang.IllegalStateException: This mutex is not locked...
- *  Others were having issues on Multiplatform Git Repo, but no solution. I just tried this one on a
- *  whim, which is suppose to prevent multiple clicks, and so far so good!
- */
-fun Modifier.clickableSingle(
-    enabled: Boolean = true,
-    onClickLabel: String? = null,
-    role: Role? = null,
-    onClick: () -> Unit
-) = this.then(composed(
-    inspectorInfo = debugInspectorInfo {
-        name = "clickable"
-        properties["enabled"] = enabled
-        properties["onClickLabel"] = onClickLabel
-        properties["role"] = role
-        properties["onClick"] = onClick
-    }
-) {
-    val multipleEventsCutter = remember { MultipleEventsCutter.get() }
-    Modifier.clickable(
-        enabled = enabled,
-        onClickLabel = onClickLabel,
-        onClick = { multipleEventsCutter.processEvent { onClick() } },
-        role = role,
-        indication = LocalIndication.current,
-        interactionSource = remember { MutableInteractionSource() }
-    )
-})
