--- app/src/main/java/com/heyzeusv/solitaire/ui/autosizetext/ImmutableWrapper.kt
@@ -1,11 +1,0 @@
-package com.heyzeusv.solitaire.ui.autosizetext
-
-import androidx.compose.runtime.Immutable
-import kotlin.reflect.KProperty
-
-@Immutable
-data class ImmutableWrapper<T>(val value: T)
-
-fun <T> T.toImmutableWrapper() = ImmutableWrapper(this)
-
-operator fun <T> ImmutableWrapper<T>.getValue(thisRef: Any?, property: KProperty<*>) = value
