+++ app/src/main/java/com/heyzeusv/solitaire/ui/game/Tableau.kt
@@ -1,6 +1,7 @@
 package com.heyzeusv.solitaire.ui.game
 
 import androidx.compose.foundation.Image
+import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.height
@@ -21,7 +22,6 @@ import com.heyzeusv.solitaire.data.Card
 import com.heyzeusv.solitaire.util.MoveResult
 import com.heyzeusv.solitaire.util.SolitairePreview
 import com.heyzeusv.solitaire.util.Suits
-import com.heyzeusv.solitaire.ui.clickableSingle
 
 /**
  *  Composable that displays Tableau [pile] with [tableauIndex]. [cardHeight] is used to shift each
@@ -54,7 +54,7 @@ fun SolitaireTableau(
                     modifier = Modifier
                         .height(cardHeight)
                         .clip(RoundedCornerShape(4.dp)) // makes click surface have round edges
-                        .clickableSingle { handleMoveResult(onClick(tableauIndex, cardIndex)) },
+                        .clickable { handleMoveResult(onClick(tableauIndex, cardIndex)) },
                     card = card
                 )
             }
