+++ app/src/main/java/com/heyzeusv/solitaire/ui/scoreboard/Scoreboard.kt
@@ -17,7 +17,7 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.heyzeusv.solitaire.R
 import com.heyzeusv.solitaire.util.SolitairePreview
-import com.heyzeusv.solitaire.ui.autosizetext.AutoSizeText
+import com.heyzeusv.solitaire.ui.AutoSizeText
 import com.heyzeusv.solitaire.util.formatTimeDisplay
 
 /**
