+++ app/src/main/java/com/heyzeusv/solitaire/util/PreviewUtil.kt
@@ -4,8 +4,8 @@ import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import com.heyzeusv.solitaire.ui.theme.PreviewBG
-import com.heyzeusv.solitaire.ui.theme.SolitaireTheme
+import com.heyzeusv.solitaire.util.theme.PreviewBG
+import com.heyzeusv.solitaire.util.theme.SolitaireTheme
 
 /**
  *  Helper function for Composable Previews. Gives semi-transparent background to previews to make
