+++ app/src/main/java/com/heyzeusv/solitaire/util/theme/Color.kt
@@ -1,4 +1,4 @@
-package com.heyzeusv.solitaire.ui.theme
+package com.heyzeusv.solitaire.util.theme
 
 import androidx.compose.ui.graphics.Color
